feat(backend): rename admin to user
