+++ backend/app.py
@@ -1,8 +1,8 @@
 from auth.jwt_bearer import JWTBearer
 from config.config import initiate_database
 from fastapi import Depends, FastAPI
-from routes.admin import router as AdminRouter
 from routes.student import router as StudentRouter
+from routes.user import router as UserRouter
 
 app = FastAPI()
 
@@ -19,7 +19,7 @@ async def read_root():
     return {"message": "Welcome to this fantastic app!"}
 
 
-app.include_router(AdminRouter, tags=["Administrator"], prefix="/admin")
+app.include_router(UserRouter, tags=["User"], prefix="/user")
 app.include_router(
     StudentRouter,
     tags=["Students"],
