+++ backend/models/__init__.py
@@ -1,4 +1,4 @@
-from models.admin import Admin
 from models.student import Student
+from models.user import User
 
-__all__ = [Student, Admin]
+__all__ = [Student, User]
