+++ backend/schemas/user.py
@@ -1,34 +1,15 @@
-from beanie import Document
 from fastapi.security import HTTPBasicCredentials
 from pydantic import BaseModel, EmailStr
 
 
-class Admin(Document):
-    fullname: str
-    email: EmailStr
-    password: str
-
-    class Config:
-        json_schema_extra = {
-            "example": {
-                "fullname": "Abdulazeez Abdulazeez Adeshina",
-                "email": "abdul@youngest.dev",
-                "password": "3xt3m#",
-            }
-        }
-
-    class Settings:
-        name = "admin"
-
-
-class AdminSignIn(HTTPBasicCredentials):
+class UserSignIn(HTTPBasicCredentials):
     class Config:
         json_schema_extra = {
             "example": {"username": "abdul@youngest.dev", "password": "3xt3m#"}
         }
 
 
-class AdminData(BaseModel):
+class UserData(BaseModel):
     fullname: str
     email: EmailStr
